VatSpy, GeoPoint2D: tolerate non-compliant coordinates
VatSpy data may not exceed +- 90 degrees on latitudes and +- 180 degrees
on longitudes. GeoPoint2D now has a "normalize" and a "wrap" construction
method to interpret such coordinates, if needed. In current VATSIM VatSpy
files, the International Date Line crosses -180 degrees longitude which
can be handled through normalization without any expected side effects.

Lines in FIR boundaries however are known to occasionally also contain
latitudes exceeding +- 90 degrees meaning they cross the poles and wrap to
the other side of the planet. As this is a common issue, such data is now
accepted but logged.